third_party/blink/web_tests/platform/mac-mac14-arm64/fast/webgl/webgl-exceptions-expected.txt

CONSOLE ERROR: Uncaught Unable to fetch WebGL rendering context for Canvas
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
